requirements.txt
#####################
# pip install torch==1.12.1+cu113 torchvision==0.13.1+cu113 torchaudio==0.12.1 --extra-index-url https://download.pytorch.org/whl/cu113
# pip install -r requirements.txt
#####################

einops
ipykernel
isort
lightning-utilities
matplotlib
numpy
pandas
patool
reformer_pytorch==1.4.4
scikit_learn
scipy
seaborn
setproctitle
sktime
sympy
tensorboard==2.12.0
torchmetrics
tqdm
xlrd
statsmodels
huggingface_hub
multiprocess
transformers
datasets
linear-attention-transformer
pydantic
docstring-parser
local-attention